+++ src/arvdevice.h
@@ -76,10 +76,10 @@ struct _ArvDeviceClass {
 	const char *	(*get_genicam_xml)	(ArvDevice *device, size_t *size);
 	ArvGc *		(*get_genicam)		(ArvDevice *device);
 
-	gboolean	(*read_memory)		(ArvDevice *device, guint32 address, guint32 size, void *buffer, GError **error);
-	gboolean	(*write_memory)		(ArvDevice *device, guint32 address, guint32 size, void *buffer, GError **error);
-	gboolean	(*read_register)	(ArvDevice *device, guint32 address, guint32 *value, GError **error);
-	gboolean	(*write_register)	(ArvDevice *device, guint32 address, guint32 value, GError **error);
+	gboolean	(*read_memory)		(ArvDevice *device, guint64 address, guint32 size, void *buffer, GError **error);
+	gboolean	(*write_memory)		(ArvDevice *device, guint64 address, guint32 size, void *buffer, GError **error);
+	gboolean	(*read_register)	(ArvDevice *device, guint64 address, guint32 *value, GError **error);
+	gboolean	(*write_register)	(ArvDevice *device, guint64 address, guint32 value, GError **error);
 
 	/* signals */
 	void		(*control_lost)		(ArvDevice *device);
@@ -89,10 +89,10 @@ GType arv_device_get_type (void);
 
 ArvStream *	arv_device_create_stream	(ArvDevice *device, ArvStreamCallback callback, void *user_data);
 
-gboolean	arv_device_read_memory 		(ArvDevice *device, guint32 address, guint32 size, void *buffer, GError **error);
-gboolean	arv_device_write_memory	 	(ArvDevice *device, guint32 address, guint32 size, void *buffer, GError **error);
-gboolean 	arv_device_read_register	(ArvDevice *device, guint32 address, guint32 *value, GError **error);
-gboolean	arv_device_write_register 	(ArvDevice *device, guint32 address, guint32 value, GError **error);
+gboolean	arv_device_read_memory 		(ArvDevice *device, guint64 address, guint32 size, void *buffer, GError **error);
+gboolean	arv_device_write_memory	 	(ArvDevice *device, guint64 address, guint32 size, void *buffer, GError **error);
+gboolean 	arv_device_read_register	(ArvDevice *device, guint64 address, guint32 *value, GError **error);
+gboolean	arv_device_write_register 	(ArvDevice *device, guint64 address, guint32 value, GError **error);
 
 const char * 	arv_device_get_genicam_xml 		(ArvDevice *device, size_t *size);
 ArvGc *		arv_device_get_genicam			(ArvDevice *device);
