+++ src/arvuvcp.h
@@ -228,9 +228,9 @@ arv_uvcp_manifest_entry_get_schema_type (ArvUvcpManifestEntry *entry)
 #undef ARAVIS_PACKED_STRUCTURE
 
 void 			arv_uvcp_packet_free 			(ArvUvcpPacket *packet);
-ArvUvcpPacket * 	arv_uvcp_packet_new_read_memory_cmd 	(guint32 address, guint32 size,
+ArvUvcpPacket * 	arv_uvcp_packet_new_read_memory_cmd 	(guint64 address, guint32 size,
 								 guint16 packet_id, size_t *packet_size);
-ArvUvcpPacket * 	arv_uvcp_packet_new_write_memory_cmd	(guint32 address, guint32 size,
+ArvUvcpPacket * 	arv_uvcp_packet_new_write_memory_cmd	(guint64 address, guint32 size,
 								 guint16 packet_id, size_t *packet_size);
 char * 			arv_uvcp_packet_to_string 		(const ArvUvcpPacket *packet);
 void 			arv_uvcp_packet_debug 			(const ArvUvcpPacket *packet, ArvDebugLevel level);
